+++ lib/bundler/downloadfile.rb
@@ -1,3 +1,24 @@
+# Copyright (c) 2020 Andy Maleh
+# 
+# Permission is hereby granted, free of charge, to any person obtaining
+# a copy of this software and associated documentation files (the
+# "Software"), to deal in the Software without restriction, including
+# without limitation the rights to use, copy, modify, merge, publish,
+# distribute, sublicense, and/or sell copies of the Software, and to
+# permit persons to whom the Software is furnished to do so, subject to
+# the following conditions:
+# 
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+# 
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+# LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+# WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
 module Bundler
   class Downloadfile    
     attr_reader :file_content, :gem_path, :keep_existing
@@ -6,10 +27,11 @@ module Bundler
     PLATFORM_OS = SUPPORTED_OPERATING_SYSTEMS.detect {|os| OS.send("#{os}?")}
     SUBCOMMANDS = %w[start clear clean]
     
-    def initialize(file_content, gem_path:, keep_existing: nil)
+    def initialize(file_content, gem_path:, keep_existing: nil, all_operating_systems: nil)
       @file_content = file_content
       @gem_path = gem_path
       @keep_existing = keep_existing
+      @all_operating_systems = all_operating_systems
       @downloads = []
       interpret
     end
@@ -32,7 +54,7 @@ module Bundler
     end
     
     def start
-      @downloads.select {|download| download.os.nil? || download.os == PLATFORM_OS }.each(&:start)
+      @downloads.select {|download| @all_operating_systems || download.os.nil? || download.os == PLATFORM_OS }.each(&:start)
     end
     
     def clear