+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add a post install message
 - `bundle download clear` command to remove downloads (alias as `clean` too)
 - Support `os` option for `download` DSL keyword in `Downloadfile`
+- `bundle download --all-operating-systems` to download files for all OS'es
 
 ## 1.0.0